# feat: Phase 1 Complete - Core Infrastructure with Critical Fixes