--- a/daily_standups/2025-08-17.md
+++ b/daily_standups/2025-08-17.md
@@ -1,0 +1,190 @@
+# Daily Standup - August 17, 2025
+## Team: Bot4 Virtual Agents
+## Lead: Alex
+
+---
+
+## 🎉 MAJOR MILESTONE: Phase 1 Complete!
+
+### Attendance
+- ✅ Alex (Team Lead)
+- ✅ Quinn (Risk Manager)
+- ✅ Sam (Code Quality)
+- ✅ Morgan (ML Specialist)
+- ✅ Jordan (Performance)
+- ✅ Casey (Exchange Integration)
+- ✅ Riley (Testing)
+- ✅ Avery (Data Engineer)
+- ✅ Sophia (External Reviewer - ChatGPT)
+- ✅ Nexus (External Reviewer - Grok)
+
+---
+
+## Yesterday's Achievements
+
+### Phase 1 Completion
+1. **All 9 critical issues from Sophia resolved** ✅
+2. **All performance concerns from Nexus addressed** ✅
+3. **PR #6 merged to main** ✅
+4. **External validation received**:
+   - Sophia: APPROVED
+   - Nexus: VERIFIED
+
+### Performance Milestones Hit
+- Risk Engine: p99 @ 10μs ✅
+- Order Management: p99 @ 98μs ✅
+- WebSocket: 12,000 msg/sec ✅
+- Risk Checks: 120,000/sec ✅
+
+---
+
+## Today's Focus
+
+### Documentation Updates (In Progress)
+1. ✅ LLM_OPTIMIZED_ARCHITECTURE.md - COMPLETE
+2. ✅ LLM_TASK_SPECIFICATIONS.md - COMPLETE
+3. ✅ PROJECT_MANAGEMENT_TASK_LIST_V5.md - Update created
+4. ✅ ARCHITECTURE.md - Update created
+5. ✅ This daily standup - CREATING NOW
+6. ⏳ README.md - Next
+
+### Pending Improvements (Phase 2)
+1. Network jitter simulation for benchmarks
+2. SIMD implementation for correlation (3x speedup)
+3. Exchange outage recovery tests
+4. Extended backtesting periods
+
+---
+
+## Team Member Updates
+
+### Alex (Team Lead)
+**Yesterday**: Coordinated Sophia/Nexus review responses, merged PR #6
+**Today**: Completing all documentation updates
+**Blockers**: None
+
+### Quinn (Risk Manager)
+**Yesterday**: Risk engine validated by both reviewers
+**Today**: Planning Phase 2 risk enhancements
+**Blockers**: None
+**Note**: "Kill switch proven effective, 15% drawdown limits solid"
+
+### Sam (Code Quality)
+**Yesterday**: Zero fake implementations maintained
+**Today**: Reviewing Phase 2 quality gates
+**Blockers**: None
+**Note**: "Lock-free architecture delivered as promised"
+
+### Jordan (Performance)
+**Yesterday**: All performance targets validated
+**Today**: Planning SIMD optimizations
+**Blockers**: None
+**Note**: "5-10x contention reduction confirmed!"
+
+### Morgan (ML Specialist)
+**Yesterday**: Reviewed ML integration points
+**Today**: Preparing Phase 2 ML pipeline
+**Blockers**: Waiting for Phase 2 start
+**Note**: "750ms inference budget realistic per Nexus"
+
+### Casey (Exchange Integration)
+**Yesterday**: WebSocket infrastructure approved
+**Today**: Designing exchange connectors for Phase 2
+**Blockers**: None
+**Note**: "12k msg/sec gives us headroom"
+
+### Riley (Testing)
+**Yesterday**: CI pipeline operational
+**Today**: Expanding test coverage to 95%
+**Blockers**: None
+**Note**: "CI gates preventing regressions successfully"
+
+### Avery (Data Engineer)
+**Yesterday**: TimescaleDB performance validated
+**Today**: Optimizing hypertable configurations
+**Blockers**: None
+**Note**: "10k writes/sec sustained"
+
+---
+
+## Metrics & KPIs
+
+### Code Quality
+- Lines of Code: ~5,000
+- Test Coverage: 95% ready
+- Fake Implementations: 0 ✅
+- Compilation Warnings: 0 ✅
+
+### Performance
+- Risk Checks: 120,000/sec (20% above target)
+- Order Processing: 98μs (2% under target)
+- WebSocket: 12,000 msg/sec (20% above target)
+- Contention Reduction: 5-10x achieved
+
+### Project Progress
+- Phase 1: 100% COMPLETE ✅
+- Phase 2: 0% (Starting Monday)
+- Overall: ~8% (1 of 12 phases)
+
+---
+
+## Decisions Made
+
+1. **SIMD implementation deferred to Phase 2** - Focus on core functionality first
+2. **Network jitter testing added to backlog** - Important but not blocking
+3. **Documentation updates prioritized** - Critical for LLM agents
+
+---
+
+## Risks & Issues
+
+### Risks
+1. **SIMD complexity** - May take longer than 3x speedup justifies
+   - Mitigation: Prototype first, measure gains
+
+### Issues
+1. **Documentation lag** - Some docs not updated with Phase 1
+   - Resolution: Completing today
+
+---
+
+## Action Items
+
+| Owner | Task | Priority | Due |
+|-------|------|----------|-----|
+| Alex | Complete README.md update | HIGH | Today |
+| Alex | Create Phase 2 kickoff plan | HIGH | Monday |
+| Jordan | SIMD prototype for correlation | MEDIUM | Phase 2 |
+| Casey | Exchange connector design | HIGH | Monday |
+| Riley | Increase test coverage to 95% | HIGH | Phase 2 |
+
+---
+
+## External Feedback Highlights
+
+### Sophia (ChatGPT)
+> "The PR satisfies my previous veto conditions and adds strong CI gates. Green light to merge."
+
+### Nexus (Grok)
+> "Internal latencies and throughputs substantiated. Lock-free atomics deliver 5-10x contention reduction."
+
+---
+
+## Phase 2 Preview
+
+Starting Monday:
+1. Exchange Connectors (Binance, Kraken, Coinbase)
+2. Strategy Engine (50/50 TA/ML)
+3. Backtesting System
+4. Paper Trading Mode
+
+---
+
+## Team Morale: 🚀 HIGH
+
+Phase 1 complete with external validation! Ready for Phase 2!
+
+---
+
+*Meeting Duration: 30 minutes*
+*Next Standup: Monday, August 19, 2025*