… docs/ARCHITECTURE_PHASE1_UPDATE.md
… docs/BACKTEST_DATA_REQUIREMENTS.md
… docs/LLM_DOCS_COMPLIANCE_REPORT.md
… docs/PHASE1_V5_UPDATE.md
… rust_core/src/tests/exchange_outage_recovery.rs